requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
transformers==4.36.2
torch==2.1.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu
pdfminer.six==20231228
rake-nltk==1.0.6
nltk==3.8.1
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.0
